Redid project section, broke out into various components
